requirements.txt
fastapi>=0.111.0
uvicorn[standard]>=0.30.0
anthropic>=0.42.0
python-dotenv>=1.0.1
gspread>=6.0.0
google-auth>=2.0.0
google-api-python-client>=2.0.0
pdfplumber>=0.10.0
httpx>=0.27.0
